requirements.txt
GnuPGInterface==0.3.2
PyAudio==0.2.4
argparse==1.2.1
chardet==2.0.1
command-not-found==0.2.44
defer==1.0.2
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
httplib2==0.7.2
jockey==0.9.7
keyring==0.7.1
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
lxkeymap==0.1
nvidia-common==0.0.0
oauth==1.0.1
pycups==1.9.61
pycurl==7.19.0
pyserial==2.7
pysmbc==1.0.13
pysqlite==2.6.3
python-apt==0.8.3ubuntu7
python-debian==0.1.21ubuntu1
pyxdg==0.19
screen-resolution-extra==0.0.0
simplejson==2.3.2
ufw==0.31.1-1
unattended-upgrades==0.1
wadllib==1.3.0
wsgiref==0.1.2
xkit==0.0.0
zope.interface==3.6.1
